requirements.txt
streamlit
google-genai
fpdf2
pillow